% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Data import}
\usage{
getdata()
}
\arguments{
\item{GISData}{Tables of gradient polygons}
}
\value{
list of gradient tables in global environment
}
\description{
Imports data to be used in GIS_Tool functions
}
