% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistoolHF_MM.R
\name{gistoolHF_MM}
\alias{gistoolHF_MM}
\title{Gradient analysis of Horsens Fjord}
\usage{
gistoolHF_MM()
}
\arguments{
\item{WBData}{Tables of gradient polygons}
}
\value{
Vertical profiles and Monod growth kinetics in various plots
}
\description{
Uses Model data with six gradient polygons and tree different scenarios
}
