% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistoolOF_MM.R
\name{gistoolOF_MM}
\alias{gistoolOF_MM}
\title{Gradient analysis of Odense Fjord}
\usage{
gistoolOF_MM()
}
\arguments{
\item{WBData}{Tables of gradient polygons}
}
\value{
Vertical profiles and Monod growth kinetics in various plots
}
\description{
Uses Model data with four gradient polygons and tree different scenarios
}
