% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistoolRF_MM.R
\name{gistoolRF_MM}
\alias{gistoolRF_MM}
\title{Gradient analysis of Roskilde Fjord}
\usage{
gistoolRF_MM()
}
\arguments{
\item{WBData}{Tables of gradient polygons}
}
\value{
Vertical profiles and Monod growth kinetics in various plots
}
\description{
Uses Model data with four gradient polygons and tree different scenarios
}
