% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistoolOF.R, R/gistoolRF.R, R/gistoolS3G4.R
\name{gistoolS3G4}
\alias{gistoolS3G4}
\title{Gradient analysis of WB (water body)}
\usage{
gistoolS3G4()

gistoolS3G4()

gistoolS3G4()
}
\arguments{
\item{WBData}{Tables of gradient polygons}
}
\value{
Vertical profiles and Monod growth kinetics in various plots

Vertical profiles and Monod growth kinetics in various plots

Vertical profiles and Monod growth kinetics in various plots
}
\description{
Uses Model data with four gradient polygons and tree different scenarios

Uses Model data with four gradient polygons and tree different scenarios

Uses Model data with four gradient polygons and tree different scenarios
}
