% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistoolS3G6.R
\name{gistoolS3G6}
\alias{gistoolS3G6}
\title{Gradient analysis of WB (water body)}
\usage{
gistoolS3G6()
}
\arguments{
\item{WBData}{Tables of gradient polygons}
}
\value{
Vertical profiles and Monod growth kinetics in various plots
}
\description{
Uses Model data with six gradient polygons and tree different scenarios
}
