% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistoolVF.R, R/gistoolVF_MM.R
\name{gistoolVF}
\alias{gistoolVF}
\title{Gradient analysis of Vejle Fjord}
\usage{
gistoolVF()

gistoolVF()
}
\arguments{
\item{WBData}{Tables of gradient polygons}
}
\value{
Vertical profiles and Monod growth kinetics in various plots

Vertical profiles and Monod growth kinetics in various plots
}
\description{
Uses Model data with five gradient polygons and tree different scenarios

Uses Model data with five gradient polygons and tree different scenarios
}
