% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistoolVF_MM.R
\name{gistoolVF_MM}
\alias{gistoolVF_MM}
\title{Gradient analysis of Vejle Fjord}
\usage{
gistoolVF_MM()
}
\arguments{
\item{WBData}{Tables of gradient polygons}
}
\value{
Vertical profiles and Monod growth kinetics in various plots
}
\description{
Uses Model data with five gradient polygons and tree different scenarios
}
